refactor(fiatTransaction): migrate fiat transaction logic to FiatTransactionService and remove legacy repository functions
