+++ web-app/src/lib/db/repositories/index.ts
@@ -2,7 +2,6 @@ export * from "./agent";
 export * from "./agentList";
 export * from "./creditCost";
 export * from "./creditTransaction";
-export * from "./fiatTransaction";
 export * from "./invitation";
 export * from "./job";
 export { default as prisma } from "./prisma";
